Refactor Controls and Steps components to remove unused props and integrate notification handling into a custom hook
