+++ src/components/Controls.tsx
@@ -1,53 +1,18 @@
 import React from 'react';
-import { Box, Button, IconButton, useMediaQuery } from '@mui/material';
-import { VolumeOff, Vibration, VolumeUp, Man, Woman, PlayArrow, Pause, Replay } from '@mui/icons-material';
-import { NotificationMode, TranslationType } from '../types';
+import { Box, Button } from '@mui/material';
+import { PlayArrow, Pause, Replay } from '@mui/icons-material';
+import { TranslationType } from '../types';
 
 interface ControlsProps {
   t: TranslationType;
   onPlay: () => void;
   onPause: () => void;
   onReset: () => void;
-  notificationMode: NotificationMode;
-  setNotificationMode: (notificationMode: NotificationMode) => void;
-  voice: 'male' | 'female';
-  setVoice: (newVoice: 'male' | 'female') => void;
 }
 
-const Controls: React.FC<ControlsProps> = ({ t, onPlay, onPause, onReset, notificationMode, setNotificationMode, voice, setVoice }) => {
-  const isSmallScreen = useMediaQuery('(max-width:465px)');
-
-  const handleToggleNotificationMode = () => {
-    const getNextMode = (currentMode: NotificationMode) => {
-      switch (currentMode) {
-        case 'none':
-          return 'vibrate';
-        case 'vibrate':
-          return 'sound';
-        case 'sound':
-          return 'none';
-      }
-    };
-    const next = getNextMode(notificationMode);
-    setNotificationMode(next);
-  };
-
-  const handleToggleVoice = () => {
-    setVoice(voice === 'male' ? 'female' : 'male');
-  };
-
+const Controls: React.FC<ControlsProps> = ({ t, onPlay, onPause, onReset }) => {
   return (
-    <Box sx={{ display: 'flex', justifyContent: 'center', mb: 5, ml: -6}}>
-      <IconButton
-        onClick={handleToggleVoice}
-        sx={{ visibility: notificationMode === 'sound' ? 'visible' : 'hidden' }}
-      >
-        {voice === 'male' ? <Man /> : <Woman />}
-      </IconButton>
-      <IconButton onClick={handleToggleNotificationMode} sx={{ mr: 1 }}>
-        {notificationMode === 'none' ? (<VolumeOff />) : notificationMode === 'vibrate' ? (<Vibration />) : (<VolumeUp />)}  
-      </IconButton>
-
+    <Box sx={{ display: 'flex', justifyContent: 'center', mb: 5 }}>
       <Button
         variant="contained"
         color="primary"
@@ -55,7 +20,7 @@ const Controls: React.FC<ControlsProps> = ({ t, onPlay, onPause, onReset, notifi
         sx={{ mr: 1, display: 'flex', alignItems: 'center', justifyContent: 'center' }}
       >
         <PlayArrow />
-        {!isSmallScreen && t.play}
+        {t.play}
       </Button>
       <Button
         variant="contained"
@@ -64,7 +29,7 @@ const Controls: React.FC<ControlsProps> = ({ t, onPlay, onPause, onReset, notifi
         sx={{ mr: 1, display: 'flex', alignItems: 'center', justifyContent: 'center' }}
       >
         <Pause />
-        {!isSmallScreen && t.pause}
+        {t.pause}
       </Button>
       <Button
         variant="contained"
@@ -73,7 +38,7 @@ const Controls: React.FC<ControlsProps> = ({ t, onPlay, onPause, onReset, notifi
         sx={{ display: 'flex', alignItems: 'center', justifyContent: 'center' }}
       >
         <Replay />
-        {!isSmallScreen && t.reset}
+        {t.reset}
       </Button>
     </Box>
   );
